Extracted kindergeld function.
