+++ src/analysis/tax_transfer_funcs/test_tax_transfers/test_kiz.py
@@ -3,8 +3,8 @@ import pytest
 from pandas.testing import assert_frame_equal
 
 from src.analysis.tax_transfer_funcs.benefits.kiz import kiz
-from src.analysis.tax_transfer_funcs.taxes import kg_eligibility_hours
-from src.analysis.tax_transfer_funcs.taxes import kg_eligibility_wage
+from src.analysis.tax_transfer_funcs.taxes.kindergeld import kg_eligibility_hours
+from src.analysis.tax_transfer_funcs.taxes.kindergeld import kg_eligibility_wage
 from src.analysis.tax_transfer_funcs.test_tax_transfers.auxiliary_test_tax import (
     load_input,
 )
